fix(ai): run Qwen/SmolVLM VLM generation in FP16 and give MIOpen writable caches
On the gfx1150 iGPU, MIOpen cannot compile the vision tower's BF16 3D-conv
kernel, so Qwen2.5-VL generation failed with miopenStatusUnknownError after
loading. Loading the VLMs in FP16 (the same dtype SigLIP embeddings already
use) lets MIOpen build the kernel; verified generation works end-to-end on
prod (~12s/caption).

Also pre-create /home/ai_user/.cache and .config owned by ai_user in the AI
image so MIOpen can write its find-db and kernel cache on every fresh
container (the model-cache volume mount leaves .cache root-owned otherwise,
which broke MIOpen kernel caching in prod).

+++ ai/ai_service.py
@@ -133,7 +133,7 @@ def load_models():
         from transformers import Qwen2_5_VLForConditionalGeneration
         vlm_model = Qwen2_5_VLForConditionalGeneration.from_pretrained(
             vlm_model_name,
-            torch_dtype=torch.bfloat16 if device == "cuda" else torch.float32,
+            torch_dtype=torch.float16 if device == "cuda" else torch.float32,
             attn_implementation="sdpa",
         ).to(device)
         # Cap visual tokens aggressively: 256 tokens max (~200K pixels)
@@ -177,7 +177,7 @@ def load_models():
                 smolvlm_model = SmolVLMForConditionalGeneration.from_pretrained(
                     smolvlm_model_name,
                     config=smolvlm_config,
-                    torch_dtype=torch.bfloat16 if device == "cuda" else torch.float32,
+                    torch_dtype=torch.float16 if device == "cuda" else torch.float32,
                 ).to(device)
                 last_exc = None
                 break
